Parser refactor.
Split the Parser.hs into multiple smaller modules. All of them are
submodules of Parser module. This change will hopefully make it easier
to navigate the code and shorten compilation times for small changes
in the parser.
