+++ src/Main.hs
@@ -4,6 +4,7 @@ import GetOpt
 import Parser
 import Text.Megaparsec (errorBundlePretty)
 import qualified Data.Text.IO as T
+import qualified Data.Text as T
 import PrettyPrint (prettyPrint)
