+++ test/parsing/Foo.zo
@@ -10,4 +10,4 @@ let _ :=
         case _       => '1'
         case x       => 123
         case []      => "123"
-        case (x::xs) => []
+        case (x::y::xs) => []
